use uin64 to replace uint32 of sequence

+++ data/inner.go
@@ -21,7 +21,7 @@ type CompressedNodeEntry struct {
 func (n InnerNode) GetType() string    { return nodeTypes[n.Type] }
 func (n InnerNode) Prefix() HashPrefix { return HP_INNER_NODE }
 func (n InnerNode) NodeType() NodeType { return n.Type }
-func (n InnerNode) Ledger() uint32     { return 0 }
+func (n InnerNode) Ledger() uint64     { return 0 }
 func (n InnerNode) GetHash() *Hash256  { return &n.Id }
 func (n InnerNode) NodeId() *Hash256   { return &n.Id }
